requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
streamlit==1.38.0
sqlalchemy==2.0.35
requests==2.32.3
python-dotenv==1.0.1
schedule==1.2.2
pandas==2.2.3
reportlab==4.2.2
